requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
psycopg==3.2.10
psycopg-binary==3.2.10
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
